test: reorganize DI tests to integration folder
Move DI tests to integration folder where they belong.
These tests verify the integration of the DI system with
the application, not unit testing of individual components.
